Add files via upload
This is a notebook based on Google Colaboratory. It contains Python programming codes for data processing and model training in AutoML.

For this study, the metallurgical industry provided data on the order of 1,000,000 records from three different machines containing information from seven sensors, which indicate specific failure modes, installed in heat treatment furnaces for metal parts, namely: global vibration, collision, imbalance, network, bearing 1, bearing 2, bearing 3.

The dataset represents an operational interval of 8 days. The set of sensors is used to monitor the vibration of the equipment during its use. For each sensor, the company establishes, based on the knowledge of its experts, value ranges, which represent intervals considered acceptable or normal.