Add files via upload
This dataset was used to implement a failure criticality classification model used in Case Study 1. The objective is to automatically classify failures as critical or not, according to supervised learning. 

Using the Scikit-Learn library, the dataset with 10,000 failure records was organized into two parts, one with 75% of the data for model training and the remaining 25% for evaluating the model's accuracy in class predictions.